+++ starkinfra/__init__.py
@@ -1,4 +1,4 @@
-version = "0.5.0"
+version = "0.6.0"
 language = "en-US"
 timeout = 15
 user = None